Move the transfer to spouse tests into their own module
There are thirty of them now, which is half of test_calc.py and more than
anything else in there is about.
