+++ tests/general/test_transfer_to_spouse.py
@@ -1,0 +1,783 @@
+"""Tests for no gain/no loss transfers to a spouse or civil partner.
+
+Covers the RAW TRANSFER_TO_SPOUSE action, how a transfer is identified
+against the transferor's own acquisitions, and the broker gift rows that
+have to be classified before anything can be calculated.
+"""
+
+from __future__ import annotations
+
+import datetime
+from decimal import Decimal
+import logging
+
+import pytest
+
+from cgt_calc.currency_converter import CurrencyConverter
+from cgt_calc.current_price_fetcher import CurrentPriceFetcher
+from cgt_calc.exceptions import (
+    CalculationError,
+    InvalidTransactionError,
+    QuantityNotPositiveError,
+    UnclassifiedGiftError,
+)
+from cgt_calc.initial_prices import InitialPrices
+from cgt_calc.isin_converter import IsinConverter
+from cgt_calc.main import CapitalGainsCalculator
+from cgt_calc.model import ActionType, RuleType
+from cgt_calc.parsers.broker_registry import _transaction_sort_key
+from cgt_calc.spin_off_handler import SpinOffHandler
+
+from .calc_test_data import (
+    split_transaction,
+    transaction,
+    transfer_to_spouse_transaction,
+)
+from .test_calc import create_calculator, get_report
+
+
+def test_transfer_to_spouse_from_pool_is_no_gain_no_loss() -> None:
+    """A transfer to spouse leaves the Section 104 pool at base cost with nil gain."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(buy_day, ActionType.BUY, "BAR", 5, 5, 0, -25, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 4),
+            transfer_to_spouse_transaction(transfer_day, "BAR", 2),
+        ],
+    )
+
+    # Not a taxable disposal.
+    assert report.total_gain() == Decimal(0)
+    assert report.disposal_count == 0
+    assert report.disposal_proceeds == Decimal(0)
+
+    # Pool reduced by 4 units at the £10 average cost.
+    assert calculator.portfolio["FOO"].quantity == Decimal(6)
+    assert calculator.portfolio["FOO"].amount == Decimal(60)
+
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    assert all(e.rule_type is RuleType.TRANSFER_TO_SPOUSE for e in entries)
+    assert sum((e.quantity for e in entries), Decimal(0)) == Decimal(4)
+    assert sum((e.gain for e in entries), Decimal(0)) == Decimal(0)
+    # £40 base cost passes to the recipient (4 units * £10 average).
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(40)
+
+    # The second symbol transferred that day is handled independently.
+    bar = report.calculation_log[transfer_day]["transfer-to-spouse$BAR"]
+    assert sum((e.allowable_cost for e in bar), Decimal(0)) == Decimal(10)
+    assert calculator.portfolio["BAR"].quantity == Decimal(3)
+
+
+def test_transfer_to_spouse_matches_repurchase_within_30_days() -> None:
+    """A purchase within 30 days after a transfer is matched first (s106A).
+
+    The base cost passing to the spouse is then the cost of the re-purchased
+    shares, not the Section 104 average - the case a plain pool reduction gets
+    wrong.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    rebuy_day = datetime.date(2024, 6, 15)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 5),
+            transaction(rebuy_day, ActionType.BUY, "FOO", 5, 12, 0, -60, "GBP"),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    # Matched to the £12 re-purchase, so base cost is £60 not the £50 average.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(60)
+    # The original 10 @ £10 pool is left untouched by the transfer.
+    assert calculator.portfolio["FOO"].quantity == Decimal(10)
+    assert calculator.portfolio["FOO"].amount == Decimal(100)
+
+
+def test_transfer_to_spouse_entire_holding_empties_pool() -> None:
+    """Transferring the whole holding removes it from the portfolio with nil gain."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 10),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    assert calculator.portfolio["FOO"].quantity == Decimal(0)
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(100)
+
+
+def test_transfer_to_spouse_matches_same_day_acquisition() -> None:
+    """A purchase on the transfer day is matched first (same day rule).
+
+    The base cost passing to the spouse is then the cost of the shares bought
+    that day, not the Section 104 average.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(transfer_day, ActionType.BUY, "FOO", 3, 20, 0, -60, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 3),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    # Matched to the £20 same-day purchase, so base cost is £60 not the £10 average.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(60)
+    # The original 10 @ £10 pool is left untouched by the transfer.
+    assert calculator.portfolio["FOO"].quantity == Decimal(10)
+    assert calculator.portfolio["FOO"].amount == Decimal(100)
+
+
+@pytest.mark.parametrize(
+    "acquisition_day",
+    [
+        pytest.param(datetime.date(2024, 6, 10), id="same-day"),
+        pytest.param(datetime.date(2024, 7, 9), id="within-30-days"),
+    ],
+)
+def test_transfer_to_spouse_same_day_as_sale_is_rejected(
+    acquisition_day: datetime.date,
+) -> None:
+    """A sale and a transfer competing for the same acquisitions is unsupported."""
+    buy_day = datetime.date(2024, 6, 1)
+    event_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with pytest.raises(CalculationError, match="does not say how to split") as err:
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transaction(event_day, ActionType.SELL, "FOO", 3, 20, 0, 60, "GBP"),
+                transfer_to_spouse_transaction(event_day, "FOO", 2),
+                transaction(
+                    acquisition_day, ActionType.BUY, "FOO", 4, 30, 0, -120, "GBP"
+                ),
+            ],
+        )
+
+    # The message names the clash concretely so it can be acted on.
+    message = str(err.value)
+    assert "sold 3 units of FOO and transferred 2" in message
+    assert str(acquisition_day) in message
+
+
+def test_transfer_to_spouse_same_day_as_sale_from_pool_is_allowed() -> None:
+    """A sale and a transfer both drawing on the pool alone are unambiguous.
+
+    Neither can be identified against an acquisition, so both take the same
+    Section 104 average cost and the order between them cannot matter.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    event_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(event_day, ActionType.SELL, "FOO", 3, 20, 0, 60, "GBP"),
+            transfer_to_spouse_transaction(event_day, "FOO", 2),
+        ],
+    )
+
+    # Only the sale is taxable: £60 proceeds against a £30 allowable cost.
+    assert report.disposal_count == 1
+    assert report.disposal_proceeds == Decimal(60)
+    assert report.total_gain() == Decimal(30)
+
+    entries = report.calculation_log[event_day]["transfer-to-spouse$FOO"]
+    # The transfer takes the same £10 average, so £20 passes to the recipient.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(20)
+    assert calculator.portfolio["FOO"].quantity == Decimal(5)
+    assert calculator.portfolio["FOO"].amount == Decimal(50)
+
+
+def test_transfer_to_spouse_more_than_owned_is_rejected() -> None:
+    """Transferring more units than held is refused."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with pytest.raises(InvalidTransactionError, match="more than the available"):
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transfer_to_spouse_transaction(transfer_day, "FOO", 11),
+            ],
+        )
+
+
+def test_transfer_to_spouse_of_not_owned_symbol_is_rejected() -> None:
+    """Transferring a symbol that was never acquired is refused."""
+    calculator = create_calculator()
+    with pytest.raises(InvalidTransactionError, match="not owned symbol"):
+        get_report(
+            calculator,
+            [transfer_to_spouse_transaction(datetime.date(2024, 6, 10), "FOO", 1)],
+        )
+
+
+def test_transfer_to_spouse_warns_about_the_ignored_price(
+    caplog: pytest.LogCaptureFixture,
+) -> None:
+    """A gift has no consideration, so any price on the row is ignored."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with caplog.at_level(logging.WARNING, logger="cgt_calc.main"):
+        report = get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transaction(
+                    transfer_day,
+                    ActionType.TRANSFER_TO_SPOUSE,
+                    "FOO",
+                    4,
+                    25,
+                    0,
+                    0,
+                    "GBP",
+                ),
+            ],
+        )
+
+    assert "Ignoring the price" in caplog.text
+    # Still the pool average, not the £25 that was in the price column.
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(40)
+
+
+def test_transfer_to_spouse_fee_is_added_to_the_base_cost() -> None:
+    """A fee is an incidental cost of the disposal, so the recipient inherits it.
+
+    s58 fixes the consideration at the amount giving neither gain nor loss, and
+    that amount has to cover the fee, so the fee ends up in the base cost the
+    recipient takes on (TCGA 1992 s38(2), CG15250).
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(
+                transfer_day,
+                ActionType.TRANSFER_TO_SPOUSE,
+                "FOO",
+                4,
+                None,
+                5,
+                -5,
+                "GBP",
+            ),
+        ],
+    )
+
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    # £40 of pool cost plus the £5 fee.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(45)
+    # Still no gain to the transferor.
+    assert sum((e.gain for e in entries), Decimal(0)) == Decimal(0)
+    assert report.total_gain() == Decimal(0)
+    # The fee does not come out of the pool the transferor keeps.
+    assert calculator.portfolio["FOO"].amount == Decimal(60)
+
+
+def test_transfer_to_spouse_shown_in_report() -> None:
+    """The text report lists each transfer with the base cost passed on."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 1000, 10, 0, -10000, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 400),
+        ],
+    )
+
+    output = str(report)
+    assert "Transferred to spouse" in output
+    # Quantities are shown as entered, like the PDF, so fractions survive.
+    assert f"{transfer_day}: FOO 400 units, base cost £4,000.00" in output
+
+
+def test_unclassified_gift_is_rejected_with_instructions() -> None:
+    """A broker gift with no classification refuses and says how to fix it."""
+    buy_day = datetime.date(2024, 6, 1)
+    gift_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with pytest.raises(UnclassifiedGiftError) as err:
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+            ],
+        )
+
+    message = str(err.value)
+    # The exact RAW line to paste, so the fix is mechanical.
+    assert "2024-06-10,TRANSFER_TO_SPOUSE,FOO,4,0.00,0.00,GBP" in message
+    # Both outcomes are named, so a non-spouse gift is not quietly treated as one.
+    assert "spouse or civil partner" in message
+    assert "disposal at market value" in message
+
+
+def test_gift_matched_by_transfer_to_spouse_is_accounted_for() -> None:
+    """A matching transfer row classifies the gift, which is then not counted twice."""
+    buy_day = datetime.date(2024, 6, 1)
+    gift_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+            transfer_to_spouse_transaction(gift_day, "FOO", 4),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    # Four units left once, not twice: the broker row is redundant, not extra.
+    assert calculator.portfolio["FOO"].quantity == Decimal(6)
+    assert calculator.portfolio["FOO"].amount == Decimal(60)
+    entries = report.calculation_log[gift_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(40)
+
+
+def test_gift_with_mismatched_transfer_is_rejected() -> None:
+    """A transfer row that does not match the gift does not classify it."""
+    buy_day = datetime.date(2024, 6, 1)
+    gift_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with pytest.raises(UnclassifiedGiftError):
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+                # Same day and symbol, different quantity.
+                transfer_to_spouse_transaction(gift_day, "FOO", 3),
+            ],
+        )
+
+
+def test_transfer_to_spouse_without_quantity_is_rejected() -> None:
+    """A transfer row has to say how many units moved."""
+    buy_day = datetime.date(2024, 6, 1)
+    calculator = create_calculator()
+    with pytest.raises(QuantityNotPositiveError):
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+                transfer_to_spouse_transaction(datetime.date(2024, 6, 10), "FOO", 0),
+            ],
+        )
+
+
+def test_gift_without_quantity_is_rejected() -> None:
+    """A gift row with no quantity is refused before asking who received it."""
+    calculator = create_calculator()
+    with pytest.raises(QuantityNotPositiveError):
+        get_report(
+            calculator,
+            [
+                transaction(
+                    datetime.date(2024, 6, 10),
+                    ActionType.GIFT,
+                    "FOO",
+                    0,
+                    currency="GBP",
+                )
+            ],
+        )
+
+
+def test_transfer_to_spouse_reserves_its_share_of_a_same_day_acquisition() -> None:
+    """An earlier sale cannot bed-and-breakfast shares a later transfer will claim.
+
+    The transfer is same-day with the repurchase, and the same-day rule beats
+    bed and breakfast, so the sale has to leave those shares alone or they get
+    counted twice.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    sell_day = datetime.date(2024, 6, 10)
+    rebuy_day = datetime.date(2024, 6, 15)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(sell_day, ActionType.SELL, "FOO", 5, 20, 0, 100, "GBP"),
+            transaction(rebuy_day, ActionType.BUY, "FOO", 5, 30, 0, -150, "GBP"),
+            transfer_to_spouse_transaction(rebuy_day, "FOO", 5),
+        ],
+    )
+
+    # The sale falls through to the pool at £10 rather than matching the £30
+    # repurchase, so the gain is £100 - £50 rather than £100 - £150.
+    assert report.total_gain() == Decimal(50)
+    sale = report.calculation_log[sell_day]["sell$FOO"]
+    assert all(e.rule_type is RuleType.SECTION_104 for e in sale)
+    # The transfer takes the repurchase under the same-day rule.
+    entries = report.calculation_log[rebuy_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(150)
+    assert calculator.portfolio["FOO"].quantity == Decimal(5)
+    assert calculator.portfolio["FOO"].amount == Decimal(50)
+
+
+def test_transfer_to_spouse_logs_a_pending_spin_off() -> None:
+    """A spin-off still pending when the transfer happens is recorded.
+
+    The transfer is identified like a disposal, so it is the event that
+    applies the outstanding cost-proportion adjustment and has to report it.
+    """
+    buy_day = datetime.date(2023, 6, 1)
+    spin_off_day = datetime.date(2023, 7, 5)
+    transfer_day = datetime.date(2024, 6, 10)
+    currency_converter = CurrencyConverter(None, {})
+    spin_off_handler = SpinOffHandler()
+    spin_off_handler.cache = {"BAR": "FOO"}
+    calculator = CapitalGainsCalculator(
+        2024,
+        currency_converter,
+        IsinConverter(),
+        CurrentPriceFetcher(
+            currency_converter,
+            {},
+            {
+                "FOO": {spin_off_day: Decimal(90)},
+                "BAR": {spin_off_day: Decimal(10)},
+            },
+        ),
+        spin_off_handler,
+        InitialPrices(),
+        interest_fund_tickers=[],
+        balance_check=False,
+    )
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 100, 10, 0, -1000, "GBP"),
+            transaction(
+                spin_off_day, ActionType.SPIN_OFF, "BAR", 100, 10, 0, currency="GBP"
+            ),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 50),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    # 90/(90+10) of the £1,000 stays with FOO, so 50 of 100 units carry £450.
+    entries = report.calculation_log[transfer_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(450)
+    # The transfer applied the adjustment, so it reports the spin-off too.
+    spin_off_entries = report.calculation_log[spin_off_day]["spin-off$FOO"]
+    assert len(spin_off_entries) == 1
+    assert spin_off_entries[0].rule_type is RuleType.SPIN_OFF
+
+
+def test_transfer_to_spouse_before_the_tax_year_still_reduces_the_pool() -> None:
+    """A historical transfer shapes the pool without appearing in the report.
+
+    The whole history is walked to build the pool, but only events inside the
+    reported period belong in the calculations.
+    """
+    buy_day = datetime.date(2022, 5, 1)
+    transfer_day = datetime.date(2022, 6, 10)
+    sell_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 4),
+            transaction(sell_day, ActionType.SELL, "FOO", 6, 20, 0, 120, "GBP"),
+        ],
+    )
+
+    assert transfer_day not in report.calculation_log
+    # The pool lost 4 units at £10, so the later sale has a £60 allowable cost.
+    assert report.disposal_count == 1
+    assert report.total_gain() == Decimal(60)
+
+
+def test_disposal_is_not_matched_against_a_same_day_split() -> None:
+    """A split is not an acquisition, so it cannot be matched on the same day.
+
+    TCGA 1992 s127 treats the new shares as acquired when the original ones
+    were (CG51805). Matching against them would give the disposal their nil
+    cost and leave the whole pool cost behind on fewer shares.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    split_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            split_transaction(split_day, "FOO", 10),
+            transaction(split_day, ActionType.SELL, "FOO", 4, 12, 0, 48, "GBP"),
+        ],
+    )
+
+    entries = report.calculation_log[split_day]["sell$FOO"]
+    assert all(e.rule_type is RuleType.SECTION_104 for e in entries)
+    # 20 units hold £100, so 4 of them cost £20, not £0.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(20)
+    assert report.total_gain() == Decimal(28)
+    assert calculator.portfolio["FOO"].amount == Decimal(80)
+
+
+def test_transfer_to_spouse_is_not_matched_against_a_same_day_split() -> None:
+    """The same, for a transfer: the recipient must not inherit a nil base cost."""
+    buy_day = datetime.date(2024, 6, 1)
+    split_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            split_transaction(split_day, "FOO", 10),
+            transfer_to_spouse_transaction(split_day, "FOO", 4),
+        ],
+    )
+
+    entries = report.calculation_log[split_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(20)
+    assert calculator.portfolio["FOO"].quantity == Decimal(16)
+    assert calculator.portfolio["FOO"].amount == Decimal(80)
+
+
+def test_each_gift_needs_its_own_classification() -> None:
+    """Two gifts of the same shares on one day are not settled by one row."""
+    buy_day = datetime.date(2024, 6, 1)
+    gift_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    with pytest.raises(UnclassifiedGiftError):
+        get_report(
+            calculator,
+            [
+                transaction(buy_day, ActionType.BUY, "FOO", 20, 10, 0, -200, "GBP"),
+                transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+                transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+                transfer_to_spouse_transaction(gift_day, "FOO", 4),
+            ],
+        )
+
+
+def test_two_gifts_are_settled_by_two_classifications() -> None:
+    """Matching one row per gift accounts for both."""
+    buy_day = datetime.date(2024, 6, 1)
+    gift_day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 20, 10, 0, -200, "GBP"),
+            transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+            transaction(gift_day, ActionType.GIFT, "FOO", 4, currency="GBP"),
+            transfer_to_spouse_transaction(gift_day, "FOO", 4),
+            transfer_to_spouse_transaction(gift_day, "FOO", 4),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(0)
+    # Both transfers left the pool: 20 - 8.
+    assert calculator.portfolio["FOO"].quantity == Decimal(12)
+    assert calculator.portfolio["FOO"].amount == Decimal(120)
+
+
+def test_transfer_to_spouse_sorts_after_a_same_day_acquisition() -> None:
+    """A hand-written transfer must not be read before the buy it depends on.
+
+    Transfers come from a RAW file while the shares come from a broker export,
+    and parsers merge in registry order, so without this the transfer is
+    validated first and fails as "not owned".
+    """
+    day = datetime.date(2024, 6, 10)
+    buy = transaction(day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP")
+    transfer = transfer_to_spouse_transaction(day, "FOO", 4)
+
+    # Worst case: the transfer is read first, as a RAW file would give it.
+    report = get_report(
+        create_calculator(),
+        sorted([transfer, buy], key=_transaction_sort_key),
+    )
+
+    entries = report.calculation_log[day]["transfer-to-spouse$FOO"]
+    assert sum((e.quantity for e in entries), Decimal(0)) == Decimal(4)
+
+
+def test_management_fee_is_not_a_contended_acquisition() -> None:
+    """A fee buys no shares, so it cannot be something two disposals fight over.
+
+    Fees are recorded in the acquisition list with a cost and no quantity, so
+    a naive "has a cost" test mistakes one for a purchase.
+    """
+    buy_day = datetime.date(2024, 6, 1)
+    event_day = datetime.date(2024, 6, 10)
+    fee_day = datetime.date(2024, 6, 20)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transaction(event_day, ActionType.SELL, "FOO", 3, 20, 0, 60, "GBP"),
+            transfer_to_spouse_transaction(event_day, "FOO", 2),
+            transaction(fee_day, ActionType.FEE, "FOO", None, None, 0, -1, "GBP"),
+        ],
+    )
+
+    assert report.total_gain() == Decimal(30)
+    entries = report.calculation_log[event_day]["transfer-to-spouse$FOO"]
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(20)
+
+
+def test_fractional_transfer_keeps_its_units_in_the_report() -> None:
+    """Rounding to two places would report a real transfer as zero units."""
+    buy_day = datetime.date(2024, 6, 1)
+    transfer_day = datetime.date(2024, 6, 10)
+    report = get_report(
+        create_calculator(),
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"),
+            transfer_to_spouse_transaction(transfer_day, "FOO", 0.001),
+        ],
+    )
+
+    assert "FOO 0.001 units" in str(report)
+
+
+def test_split_shares_do_not_dilute_a_same_day_purchase() -> None:
+    """A split on the day of a purchase must not spread its cost over free shares.
+
+    Both land in the acquisition list for the day, so treating the total as one
+    acquisition prices the purchase across shares that cost nothing.
+    """
+    day = datetime.date(2024, 6, 10)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(
+                datetime.date(2024, 6, 1), ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"
+            ),
+            split_transaction(day, "FOO", 10),
+            transaction(day, ActionType.BUY, "FOO", 5, 20, 0, -100, "GBP"),
+            transfer_to_spouse_transaction(day, "FOO", 5),
+        ],
+    )
+
+    entries = report.calculation_log[day]["transfer-to-spouse$FOO"]
+    # The five shares bought that day cost £100, and that is what is passed on.
+    assert sum((e.allowable_cost for e in entries), Decimal(0)) == Decimal(100)
+    # 10 original plus 10 free shares, still holding the original £100.
+    assert calculator.portfolio["FOO"].quantity == Decimal(20)
+    assert calculator.portfolio["FOO"].amount == Decimal(100)
+
+
+def test_transfer_reservation_survives_a_split_before_the_repurchase() -> None:
+    """Reservations are counted in the acquisition's units, not the disposal's.
+
+    A split between an earlier sale and the repurchase means the two are
+    counted differently; subtracting one from the other unconverted drove the
+    available quantity negative.
+    """
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(
+                datetime.date(2024, 6, 1), ActionType.BUY, "FOO", 10, 10, 0, -100, "GBP"
+            ),
+            transaction(
+                datetime.date(2024, 6, 5), ActionType.SELL, "FOO", 4, 15, 0, 60, "GBP"
+            ),
+            split_transaction(datetime.date(2024, 6, 8), "FOO", 6),
+            transaction(
+                datetime.date(2024, 6, 12), ActionType.BUY, "FOO", 10, 5, 0, -50, "GBP"
+            ),
+            transfer_to_spouse_transaction(datetime.date(2024, 6, 12), "FOO", 6),
+        ],
+    )
+
+    assert report.disposal_count == 1
+    assert report.total_gain() == Decimal(20)
+
+
+def test_fully_matched_acquisition_is_not_contended() -> None:
+    """An acquisition its own day's sale consumes cannot reach an earlier day."""
+    buy_day = datetime.date(2024, 6, 1)
+    event_day = datetime.date(2024, 6, 10)
+    later = datetime.date(2024, 6, 15)
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(buy_day, ActionType.BUY, "FOO", 20, 10, 0, -200, "GBP"),
+            transaction(event_day, ActionType.SELL, "FOO", 3, 20, 0, 60, "GBP"),
+            transfer_to_spouse_transaction(event_day, "FOO", 2),
+            transaction(later, ActionType.BUY, "FOO", 4, 30, 0, -120, "GBP"),
+            transaction(later, ActionType.SELL, "FOO", 4, 31, 0, 124, "GBP"),
+        ],
+    )
+
+    # £30 on the 10th (3 units at £10 cost) plus £4 on the 15th.
+    assert report.total_gain() == Decimal(34)
+
+
+def test_acquisition_an_earlier_disposal_took_is_not_contended() -> None:
+    """A settled bed and breakfast match leaves nothing to argue over.
+
+    The 5 June sale takes the whole 15 June purchase, so neither the sale nor
+    the transfer on 10 June can reach it and both fall to the pool.
+    """
+    calculator = create_calculator()
+    report = get_report(
+        calculator,
+        [
+            transaction(
+                datetime.date(2024, 6, 1), ActionType.BUY, "FOO", 20, 10, 0, -200, "GBP"
+            ),
+            transaction(
+                datetime.date(2024, 6, 5), ActionType.SELL, "FOO", 4, 15, 0, 60, "GBP"
+            ),
+            transaction(
+                datetime.date(2024, 6, 10), ActionType.SELL, "FOO", 3, 20, 0, 60, "GBP"
+            ),
+            transfer_to_spouse_transaction(datetime.date(2024, 6, 10), "FOO", 2),
+            transaction(
+                datetime.date(2024, 6, 15), ActionType.BUY, "FOO", 4, 30, 0, -120, "GBP"
+            ),
+        ],
+    )
+
+    # £60 loss bed-and-breakfasted on 5 June, £30 gain from the pool on the 10th.
+    assert report.total_gain() == Decimal(-30)
+    entries = report.calculation_log[datetime.date(2024, 6, 10)][
+        "transfer-to-spouse$FOO"
+    ]
+    assert all(e.rule_type is RuleType.TRANSFER_TO_SPOUSE for e in entries)
